requirements/development.txt
# SHA1:80db36131ba9a8df7c34810cee7788f03cfb68b8
#
# This file is autogenerated by pip-compile-multi
# To update, run:
#
#    pip-compile-multi
#
-r base.txt
-e file:.
    # via
    #   -r requirements/base.in
    #   -r requirements/development.in
appnope==0.1.3
    # via ipython
asttokens==2.0.5
    # via stack-data
backcall==0.2.0
    # via ipython
boto3==1.18.19
    # via tabulator
botocore==1.21.19
    # via
    #   boto3
    #   s3transfer
cached-property==1.5.2
    # via tableschema
certifi==2021.10.8
    # via requests
chardet==4.0.0
    # via tabulator
charset-normalizer==2.0.12
    # via requests
decorator==5.1.1
    # via ipython
et-xmlfile==1.1.0
    # via openpyxl
executing==0.8.3
    # via stack-data
flask-cors==3.0.10
    # via apache-superset
future==0.18.2
    # via pyhive
ijson==3.1.4
    # via tabulator
ipython==8.3.0
    # via -r requirements/development.in
jedi==0.18.1
    # via ipython
jmespath==0.10.0
    # via
    #   boto3
    #   botocore
jsonlines==2.0.0
    # via tabulator
linear-tsv==1.1.0
    # via tabulator
matplotlib-inline==0.1.3
    # via ipython
mysqlclient==2.1.0
    # via apache-superset
openai
openpyxl==3.0.7
    # via tabulator
parso==0.8.3
    # via jedi
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pillow==9.3.0
    # via apache-superset
progress==1.6
    # via -r requirements/development.in
psycopg2-binary==2.9.5
    # via apache-superset
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.2
    # via stack-data
pure-sasl==0.6.2
    # via thrift-sasl
pydruid==0.6.5
    # via apache-superset
pygments==2.12.0
    # via ipython
pyhive[hive]==0.6.5
    # via apache-superset
pyinstrument==4.0.2
    # via -r requirements/development.in
requests==2.26.0
    # via
    #   pydruid
    #   tableschema
    #   tabulator
rfc3986==1.5.0
    # via tableschema
s3transfer==0.5.0
    # via boto3
sasl==0.3.1
    # via pyhive
sqloxide==0.1.17
    # via -r requirements/development.in
stack-data==0.2.0
    # via ipython
tableschema==1.20.2
    # via apache-superset
tabulator==1.53.5
    # via tableschema
thrift==0.14.1
    # via
    #   apache-superset
    #   pyhive
    #   thrift-sasl
thrift-sasl==0.4.3
    # via pyhive
traitlets==5.2.1.post0
    # via
    #   ipython
    #   matplotlib-inline
unicodecsv==0.14.1
    # via
    #   tableschema
    #   tabulator
xlrd==2.0.1
    # via tabulator
web3>=6.6.0
# The following packages are considered to be unsafe in a requirements file:
# setuptools
